+++ src/Git/Fmt.hs
@@ -22,6 +22,7 @@ module Git.Fmt (
     handle,
 ) where
 
+import           Control.Applicative
 import           Control.Concurrent
 import           Control.Monad.Catch    (MonadMask)
 import           Control.Monad.Extra
@@ -32,6 +33,7 @@ import qualified Control.Monad.Parallel as Parallel
 import           Control.Monad.Reader
 
 import           Data.List.Extra   (chunksOf, linesBy, lower, nub)
+import           Data.Maybe        (fromMaybe)
 import qualified Data.Text         as T
 import           Data.Yaml         (prettyPrintParseException)
 import           Data.Yaml.Include (decodeFileEither)
@@ -49,10 +51,11 @@ import System.IO.Temp
 
 -- | Options.
 data Options = Options {
-        optChatty :: Chatty,
-        optNull   :: Bool,
-        optMode   :: Mode,
-        argPaths  :: [FilePath]
+        optChatty     :: Chatty,
+        optNull       :: Bool,
+        optNumThreads :: Maybe Int,
+        optMode       :: Mode,
+        argPaths      :: [FilePath]
     }
     deriving (Eq, Show)
 
@@ -73,7 +76,8 @@ handle options = do
             (liftIO $ listFilesRecursive path)
             (return [path])
             )
-    numThreads  <- liftIO getNumCapabilities
+    numThreads  <- liftIO getNumCapabilities >>= \numCapabilities ->
+        return $ fromMaybe numCapabilities (optNumThreads options)
 
     unlessM (liftIO . doesFileExist $ gitDir </> Config.defaultFileName) $ panic (gitDir </> Config.defaultFileName ++ ": not found")